Improve tests (#53)
Improve tests
