+++ .php_cs.dist
@@ -1,15 +1,17 @@
 <?php
 return PhpCsFixer\Config::create()
-    ->setRules(array(
+    ->setRules([
         '@PSR2' => true,
         'linebreak_after_opening_tag' => true,
         'ordered_imports' => true,
-    ))
+        'no_unused_imports' => true,
+        'array_syntax' => [
+            'syntax' => 'short',
+        ]
+    ])
     ->setFinder(
         PhpCsFixer\Finder::create()
             ->name('*.php')
-            ->notName('*Spec.php')
-            ->notName('*Integration.php')
-            ->in(__DIR__ . '/src')
+            ->in([__DIR__ . '/src', __DIR__ . '/Tests'])
     )
 ;
