+++ src/DependencyInjection/PimExtendedAttributeTypeExtension.php
@@ -2,7 +2,6 @@
 
 namespace Pim\Bundle\ExtendedAttributeTypeBundle\DependencyInjection;
 
-use Pim\Bundle\ElasticSearchBundle\Query\ProductQueryUtility;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
